+++ src/storage/ldb/leveldb/util/coding.cc
@@ -81,7 +81,7 @@ void PutVarint32(std::string* dst, uint32_t v) {
 char* EncodeVarint64(char* dst, uint64_t v) {
   static const int B = 128;
   unsigned char* ptr = reinterpret_cast<unsigned char*>(dst);
-  while (v >= B) {
+  while (v >= (uint64_t)B) {
     *(ptr++) = (v & (B-1)) | B;
     v >>= 7;
   }
